requirements.txt
flask
pandas
numpy
nltk
faiss-cpu
scikit-learn
sentence-transformers
openpyxl